feat: global publish button
