+++ src/app.tsx
@@ -1,12 +1,16 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { history, Link } from 'umi';
 import Footer from '@/components/Footer';
 import type { RunTimeLayoutConfig } from 'umi';
 import { useModel } from '@@/plugin-model/useModel';
 import { PageLoading } from '@ant-design/pro-layout';
-import { Typography, Avatar, Card, Dropdown } from 'antd';
+import { Typography, Avatar, Card, Dropdown, Button, message, notification } from 'antd';
 import { DownOutlined, ExportOutlined } from '@ant-design/icons';
-import { fetchPostsPublished, getDefaultSiteConfig } from '@/services/api/meta-cms';
+import {
+  deployAndPublishSite,
+  fetchPostsPublished,
+  getDefaultSiteConfig,
+} from '@/services/api/meta-cms';
 import { queryCurrentUser, queryInvitations, refreshTokens } from './services/api/meta-ucenter';
 import MenuMoreInfo from './components/MenuMoreInfo';
 import MenuUserInfo from './components/MenuUserInfo';
@@ -24,7 +28,10 @@ function CustomSiderMenu({
   initialState: any;
   menuItemProps: SiderMenuProps;
 }) {
-  const { deployedSite, setDeployedSite } = useModel('storage');
+  const [publishLoading, setPublishLoading] = useState<boolean>(false);
+  const { deployedSite, setDeployedSite, siteNeedToDeploy, setSiteNeedToDeploy } =
+    useModel('storage');
+
   useEffect(() => {
     getDefaultSiteConfig().then((response) => {
       if (response.statusCode === 200) {
@@ -36,6 +43,36 @@ function CustomSiderMenu({
       }
     });
   }, [setDeployedSite]);
+
+  const publishSiteRequest = async () => {
+    const done = message.loading('开始部署站点，请稍候…', 0);
+    setPublishLoading(true);
+
+    if (deployedSite.configId) {
+      const response = await deployAndPublishSite(deployedSite.configId);
+      console.log(response);
+      if (response.statusCode === 201) {
+        notification.success({
+          message: '任务完成',
+          description: 'Meta Space 已经重新部署成功！',
+          duration: 0,
+        });
+        setSiteNeedToDeploy(false);
+      } else {
+        notification.error({
+          message: '任务失败',
+          description: 'Meta Space 部署失败 ，请重新尝试或向开发团队反馈。',
+          duration: 0,
+        });
+      }
+    } else {
+      message.error('未获取到站点信息，无法发布文章。请先创建站点。');
+    }
+
+    done();
+    setPublishLoading(false);
+  };
+
   return (
     <div className="menu-extra-cards">
       <Dropdown overlay={<MenuUserInfo />} placement="bottomCenter" trigger={['click', 'hover']}>
@@ -67,6 +104,22 @@ function CustomSiderMenu({
           </Card>
         </a>
       )}
+      <div
+        className={`global-publish-button-background ${
+          siteNeedToDeploy
+            ? 'global-publish-button-background-visible'
+            : 'global-publish-button-background-invisible'
+        }`}
+      >
+        <Button
+          key="publish-button"
+          loading={publishLoading}
+          onClick={publishSiteRequest}
+          className="global-publish-button"
+        >
+          部署我的 Meta Space
+        </Button>
+      </div>
     </div>
   );
 }
