+++ src/components/Guide/CDNSetting.tsx
@@ -1,30 +1,25 @@
 import { Alert } from 'antd';
 import { FormattedMessage } from 'umi';
-import styles from './styles.less';
 
 export default () => (
-  <div className={styles.container}>
-    <div>
-      <Alert
-        message={<FormattedMessage id="guide.cdn.message" />}
-        description={
-          <FormattedMessage id="guide.cdn.info">
-            {(msg: string) =>
-              msg
-                ?.trim()
-                .split('\n')
-                .map((e, index) => (
-                  <div key={`cdninfo${index + 1}`}>
-                    {e}
-                    <br />
-                  </div>
-                ))
-            }
-          </FormattedMessage>
+  <Alert
+    message={<FormattedMessage id="guide.cdn.message" />}
+    description={
+      <FormattedMessage id="guide.cdn.info">
+        {(msg: string) =>
+          msg
+            ?.trim()
+            .split('\n')
+            .map((e, index) => (
+              <div key={`cdninfo${index + 1}`}>
+                {e}
+                <br />
+              </div>
+            ))
         }
-        type="success"
-        showIcon
-      />
-    </div>
-  </div>
+      </FormattedMessage>
+    }
+    type="success"
+    showIcon
+  />
 );
