+++ src/components/Guide/DomainSetting.tsx
@@ -4,7 +4,6 @@ import ProForm, { ProFormText } from '@ant-design/pro-form';
 import { message } from 'antd';
 import { useState } from 'react';
 import { useIntl } from 'umi';
-import styles from './styles.less';
 
 export default () => {
   const intl = useIntl();
@@ -18,46 +17,44 @@ export default () => {
   };
 
   return (
-    <div className={styles.container}>
-      <ProForm
-        style={{ width: 500 }}
-        name="site-info"
-        initialValues={{ domain: domainSetting }}
-        onFinish={updateDomainSettings}
-        requiredMark="optional"
-      >
-        <ProFormText
-          width="md"
-          fieldProps={{
-            addonAfter: `.${META_SPACE_BASE_DOMAIN || 'metaspaces.me'}`,
-          }}
-          name="domain"
-          placeholder={intl.formatMessage({ id: 'messages.domain.enterPrefixDomain' })}
-          validateStatus={isSuccess ? 'success' : undefined}
-          help={isSuccess ? intl.formatMessage({ id: 'notifications.domain.isValid' }) : undefined}
-          rules={[
-            {
-              validator: async (_, value) => {
-                if (!value) {
-                  setIsSuccess(false);
-                  return Promise.reject(
-                    new Error(intl.formatMessage({ id: 'notifications.domain.shouldNotBeEmpty' })),
-                  );
-                }
-                const isForbidden = await isDomainForbidden(value);
-                if (isForbidden) {
-                  setIsSuccess(false);
-                  return Promise.reject(
-                    new Error(intl.formatMessage({ id: 'notifications.domain.isForbidden' })),
-                  );
-                }
-                setIsSuccess(true);
-                return Promise.resolve();
-              },
+    <ProForm
+      style={{ width: 500 }}
+      name="site-info"
+      initialValues={{ domain: domainSetting }}
+      onFinish={updateDomainSettings}
+      requiredMark="optional"
+    >
+      <ProFormText
+        width="md"
+        fieldProps={{
+          addonAfter: `.${META_SPACE_BASE_DOMAIN || 'metaspaces.me'}`,
+        }}
+        name="domain"
+        placeholder={intl.formatMessage({ id: 'messages.domain.enterPrefixDomain' })}
+        validateStatus={isSuccess ? 'success' : undefined}
+        help={isSuccess ? intl.formatMessage({ id: 'notifications.domain.isValid' }) : undefined}
+        rules={[
+          {
+            validator: async (_, value) => {
+              if (!value) {
+                setIsSuccess(false);
+                return Promise.reject(
+                  new Error(intl.formatMessage({ id: 'notifications.domain.shouldNotBeEmpty' })),
+                );
+              }
+              const isForbidden = await isDomainForbidden(value);
+              if (isForbidden) {
+                setIsSuccess(false);
+                return Promise.reject(
+                  new Error(intl.formatMessage({ id: 'notifications.domain.isForbidden' })),
+                );
+              }
+              setIsSuccess(true);
+              return Promise.resolve();
             },
-          ]}
-        />
-      </ProForm>
-    </div>
+          },
+        ]}
+      />
+    </ProForm>
   );
 };
