+++ src/components/Guide/ThemeSetting.tsx
@@ -1,8 +1,3 @@
-import styles from './styles.less';
 import ThemePicker from '@/components/ThemePicker';
 
-export default () => (
-  <div className={styles.container}>
-    <ThemePicker />
-  </div>
-);
+export default () => <ThemePicker />;
