--- src/components/Guide/styles.less
@@ -1,11 +1,0 @@
-@import '~antd/es/style/themes/default.less';
-
-:global {
-  .ant-form-item-has-success .ant-form-item-explain {
-    color: #52c41a;
-  }
-}
-
-.container {
-  // for global container styles
-}
