+++ src/components/StorePicker/index.tsx
@@ -48,7 +48,7 @@ export default () => {
             message.success(`已成功选择仓储为 ${storeConfirmed}`);
             break;
           }
-          // TODO: add 'Gitee' as a valid store option
+          // TODO: Only works for Github
           default: {
             throw new Error('未知的仓储');
           }
@@ -96,7 +96,11 @@ export default () => {
       />
       <p>
         <FormattedMessage id="guide.storage.currentStorage" />
-        <strong>{storeConfirmed || '未选择'}</strong>
+        <strong>
+          {storeConfirmed && storeSetting.username
+            ? `${storeConfirmed} (${storeSetting.username})`
+            : '未选择'}
+        </strong>
       </p>
 
       <PlatformModal
